Renamed a test file to include .spec extension and update its tests
